Set ios tokens
